+++ logging/logadmin/metrics.go
@@ -26,7 +26,7 @@ import (
 // Metric describes a logs-based metric. The value of the metric is the
 // number of log entries that match a logs filter.
 //
-// Metrics are a feature of Stackdriver Monitoring.
+// Metrics are a feature of Cloud Monitoring.
 // See https://cloud.google.com/monitoring/api/v3/metrics for more about them.
 type Metric struct {
 	// ID is a client-assigned metric identifier. Example: