+++ spanner/spansql/types.go
@@ -57,15 +57,28 @@ func (ct *CreateTable) clearOffset() {
 // TableConstraint represents a constraint on a table.
 type TableConstraint struct {
 	Name       ID // may be empty
-	ForeignKey ForeignKey
+	Constraint Constraint
 
-	Position Position // position of the "CONSTRAINT" or "FOREIGN" token
+	Position Position // position of the "CONSTRAINT" token, or Constraint.Pos()
 }
 
 func (tc TableConstraint) Pos() Position { return tc.Position }
 func (tc *TableConstraint) clearOffset() {
+	switch c := tc.Constraint.(type) {
+	case ForeignKey:
+		c.clearOffset()
+		tc.Constraint = c
+	case Check:
+		c.clearOffset()
+		tc.Constraint = c
+	}
 	tc.Position.Offset = 0
-	tc.ForeignKey.clearOffset()
+}
+
+type Constraint interface {
+	isConstraint()
+	SQL() string
+	Node
 }
 
 // Interleave represents an interleave clause of a CREATE TABLE statement.
@@ -204,7 +217,25 @@ type Delete struct {
 func (d *Delete) String() string { return fmt.Sprintf("%#v", d) }
 func (*Delete) isDMLStmt()       {}
 
-// TODO: Insert, Update.
+// TODO: Insert.
+
+// Update represents an UPDATE statement.
+// https://cloud.google.com/spanner/docs/dml-syntax#update-statement
+type Update struct {
+	Table ID
+	Items []UpdateItem
+	Where BoolExpr
+
+	// TODO: Alias
+}
+
+func (u *Update) String() string { return fmt.Sprintf("%#v", u) }
+func (*Update) isDMLStmt()       {}
+
+type UpdateItem struct {
+	Column ID
+	Value  Expr // or nil for DEFAULT
+}
 
 // ColumnDef represents a column definition as part of a CREATE TABLE
 // or ALTER TABLE statement.
@@ -243,6 +274,19 @@ type ForeignKey struct {
 
 func (fk ForeignKey) Pos() Position { return fk.Position }
 func (fk *ForeignKey) clearOffset() { fk.Position.Offset = 0 }
+func (ForeignKey) isConstraint()    {}
+
+// Check represents a check constraint as part of a CREATE TABLE
+// or ALTER TABLE statement.
+type Check struct {
+	Expr BoolExpr
+
+	Position Position // position of the "CHECK" token
+}
+
+func (c Check) Pos() Position { return c.Position }
+func (c *Check) clearOffset() { c.Position.Offset = 0 }
+func (Check) isConstraint()   {}
 
 // Type represents a column type.
 type Type struct {
@@ -391,7 +435,7 @@ type LiteralOrParam interface {
 
 type ArithOp struct {
 	Op       ArithOperator
-	LHS, RHS Expr // only RHS is set for Neg, BitNot
+	LHS, RHS Expr // only RHS is set for Neg, Plus, BitNot
 }
 
 func (ArithOp) isExpr() {}
@@ -400,6 +444,7 @@ type ArithOperator int
 
 const (
 	Neg    ArithOperator = iota // unary -
+	Plus                        // unary +
 	BitNot                      // unary ~
 	Mul                         // *
 	Div                         // /
@@ -415,7 +460,7 @@ const (
 
 type LogicalOp struct {
 	Op       LogicalOperator
-	LHS, RHS BoolExpr // only RHS is set for Neg, BitNot
+	LHS, RHS BoolExpr // only RHS is set for Not
 }
 
 func (LogicalOp) isBoolExpr() {}